fix: rewrite topic store for incremental update
Change BoltDB schema for the topic store to
allo incremental update by AlertID. Also clear
AlertID status from the topic store on reset
to OK status.

closes https://github.com/influxdata/kapacitor/issues/2723

+++ services/storage/storagetest/storage.go
@@ -1,18 +1,23 @@
 package storagetest
 
 import (
+	"fmt"
 	"os"
 	"path"
 
+	"github.com/influxdata/kapacitor/services/alert"
 	"github.com/influxdata/kapacitor/services/storage"
 	bolt "go.etcd.io/bbolt"
 )
 
 type TestStore struct {
-	db        *BoltDB
-	versions  storage.Versions
-	registrar *storage.StoreActionerRegistrar
+	db         *BoltDB
+	versions   storage.Versions
+	registrar  *storage.StoreActionerRegistrar
+	diagnostic storage.Diagnostic
 }
+
+const memoryPath = ":memory:"
 
 // BoltDB is a database that deletes itself when closed
 type BoltDB struct {
@@ -21,7 +26,7 @@ type BoltDB struct {
 
 // NewBolt is an in-memory db that deletes itself when closed, do not use except for testing.
 func NewBolt() (*BoltDB, error) {
-	db, err := bolt.Open(":memory:", 0600, &bolt.Options{
+	db, err := bolt.Open(memoryPath, 0600, &bolt.Options{
 		Timeout:    0,
 		NoGrowSync: false,
 		MemOnly:    true,
@@ -37,37 +42,71 @@ func (b BoltDB) Store(bucket string) storage.Interface {
 }
 
 func (b BoltDB) Close() error {
+	dbPath := b.Path()
 	err := b.DB.Close()
 	if err != nil {
 		return err
 	}
-	return os.RemoveAll(path.Dir(b.Path()))
+
+	if dbPath != "" && dbPath != memoryPath {
+		return os.RemoveAll(path.Dir(b.Path()))
+	} else {
+		return nil
+	}
 }
 
-func New() TestStore {
+func New(diagnostic storage.Diagnostic) *TestStore {
 	db, err := NewBolt()
 	if err != nil {
 		panic(err)
 	}
-	return TestStore{
-		db:        db,
-		versions:  storage.NewVersions(db.Store("versions")),
-		registrar: storage.NewStorageResitrar(),
+	return &TestStore{
+		db:         db,
+		versions:   storage.NewVersions(db.Store("versions")),
+		registrar:  storage.NewStorageRegistrar(),
+		diagnostic: diagnostic,
 	}
 }
 
-func (s TestStore) Store(name string) storage.Interface {
+func (s *TestStore) Store(name string) storage.Interface {
 	return s.db.Store(name)
 }
 
-func (s TestStore) Versions() storage.Versions {
+func (s *TestStore) Versions() storage.Versions {
 	return s.versions
 }
 
-func (s TestStore) Register(name string, store storage.StoreActioner) {
+func (s *TestStore) Register(name string, store storage.StoreActioner) {
 	s.registrar.Register(name, store)
 }
 
-func (s TestStore) Close() {
-	//s.db.Close()
+func (s *TestStore) Close() error {
+	return s.db.Close()
+}
+
+func (s *TestStore) Diagnostic() storage.Diagnostic {
+	return s.diagnostic
+}
+
+func (s *TestStore) BucketEntries(topic string, alertID string) (keys []string, exists bool, err error) {
+	store := s.db.Store(alert.TopicStatesNameSpace)
+	err = store.View(func(tx storage.ReadOnlyTx) error {
+		bucket := tx.Bucket([]byte(topic))
+		if bucket == nil {
+			return fmt.Errorf("%q: %w", topic, bolt.ErrBucketNotFound)
+		}
+		if kvs, err := bucket.List(""); err != nil {
+			return fmt.Errorf("failed to list contents of bucket %q: %w", topic, err)
+		} else {
+			keys = make([]string, 0, len(kvs))
+			for _, aID := range kvs {
+				keys = append(keys, aID.Key)
+				if aID.Key == alertID {
+					exists = true
+				}
+			}
+		}
+		return nil
+	})
+	return keys, exists, err
 }
